+++ database/migrations/2018_08_15_102039_move_albums.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Album;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -14,22 +15,34 @@ class MoveAlbums extends Migration
      */
     public function up()
     {
-        if(Schema::hasTable(env('DB_OLD_LYCHEE_PREFIX','').'lychee_albums')){
-            $results = DB::table(env('DB_OLD_LYCHEE_PREFIX','').'lychee_albums')->select('*')->get();
-            foreach ($results as $result)
-            {
-                $id = $result->id;
-                $id = substr($id,1,10);
-                $id[0] = strval(intval($id[0]) % 4);
-                $album = new Album();
-                $album->id = $id;
-                $album->title = $result->title;
-                $album->description = $result->description;
-                $album->public = $result->public;
-                $album->visible_hidden = $result->visible;
-                $album->save();
-            }
-        }
+    	if(count(Album::all()) == 0) {
+		    if (Schema::hasTable(env('DB_OLD_LYCHEE_PREFIX', '') . 'lychee_albums')) {
+			    $results = DB::table(env('DB_OLD_LYCHEE_PREFIX', '') . 'lychee_albums')->select('*')->get();
+			    foreach ($results as $result) {
+				    $id = $result->id;
+				    $id = substr($id, 1, 10);
+				    $id[0] = strval(intval($id[0]) % 4);
+				    $album = new Album();
+				    $album->id = $id;
+				    $album->title = $result->title;
+				    $album->description = $result->description;
+				    $album->public = $result->public;
+				    $album->visible_hidden = $result->visible;
+				    $album->license = $result->license;
+				    $album->save();
+			    }
+		    } else {
+			    echo env('DB_OLD_LYCHEE_PREFIX', '') . "lychee_albums does not exists!\n";
+		    }
+	    }
+    	else
+	    {
+		    echo "albums is not empty.\n";
+	    }
+
+
+
+	    Album::reset_takestamp();
     }
 
     /**
@@ -39,8 +52,10 @@ class MoveAlbums extends Migration
      */
     public function down()
     {
-        if(Schema::hasTable('lychee_albums')) {
-            DB::table('albums')->delete();
-        }
+	    if(env('DB_DROP_CLEAR_TABLES_ON_ROLLBACK',false)) {
+		    if (Schema::hasTable('lychee_albums')) {
+			    DB::table('albums')->delete();
+		    }
+	    }
     }
 }
